add JWT auth

+++ project/routes/web.php
@@ -19,12 +19,12 @@ $router->get('/', function () use ($router) {
 
 $router->group(['prefix' => 'api'], function () use ($router) {
     $router->group(['prefix' => 'users'], function () use ($router) {
-        $router->get('/', 'UserController@list');
-        $router->post('/register', 'UserController@register');
         $router->post('/login', 'UserController@login');
-        $router->get('/{id}', 'UserController@getByID');
-        $router->put('/{id}', 'UserController@update');
-        $router->delete('/{id}', 'UserController@delete');
+        $router->post('/register', 'UserController@register');
+        $router->get('/', ['middleware' => ['jwt.auth', 'admin'], 'uses' => 'UserController@list']);
+        $router->get('/{id}', ['middleware' => ['jwt.auth', 'admin'],  'uses' => 'UserController@getByID']);
+        $router->put('/{id}', ['middleware' => 'jwt.auth', 'uses' => 'UserController@update']);
+        $router->delete('/{id}', ['middleware' => 'jwt.auth', 'uses' => 'UserController@delete']);
     });
 
     $router->group(['prefix' => 'lottery_games'], function () use ($router) {
@@ -32,12 +32,15 @@ $router->group(['prefix' => 'api'], function () use ($router) {
     });
 
     $router->group(['prefix' => 'lottery_game_matches'], function () use ($router) {
-        $router->post('/', 'LotteryGameMatchController@create');
-        $router->put('/{id}', 'LotteryGameMatchController@finish');
         $router->get('/', 'LotteryGameMatchController@getByLotteryID');
+
+        $router->group(['middleware' => ['jwt.auth', 'admin']], function () use ($router) {
+            $router->post('/', 'LotteryGameMatchController@create');
+            $router->put('/{id}', 'LotteryGameMatchController@finish');
+        });
     });
 
-    $router->group(['prefix' => 'lottery_game_match_users'], function () use ($router) {
+    $router->group(['prefix' => 'lottery_game_match_users', 'middleware' => 'jwt.auth'], function () use ($router) {
         $router->post('/', 'LotteryGameMatchUserController@signUpForMatch');
     });
 });
